refactor how we initialize a DiffDelta using an initializer
